Cuentas bancarias, banco y Tienda

+++ PayMe.java
@@ -1,26 +1,29 @@
 /**
- * Clase para simular el sistema PayMe y poder conectar un negocio con el banco para recibir pagos a través de cuentas bancarias
+ * Clase para simular el sistema PayMe y poder conectar un negocio con el banco
+ * para recibir pagos a través de cuentas bancarias
  */
 public class PayMe {
     private Banco banco;
 
-    public PayMe(Banco banco){
-        this.banco=banco;
+    public PayMe(Banco banco) {
+        this.banco = banco;
     }
 
     /**
      * Método para saber si existe un número de cuenta registrado en el banco
+     * 
      * @return boolean true si existe la cuenta, false en otro caso
      */
-    public boolean existeCuenta(int numCuenta){
+    public boolean existeCuenta(Long numCuenta) {
         return banco.existeCuenta(numCuenta);
     }
 
     /**
      * Método para tranferir dinero de una cuenta a otra
+     * 
      * @return boolean true si la transferencia fue exitosa, false en otro caso
      */
-    public boolean transferir(int numCuentaOrigen, int cvvOrigen, int numCuentaDestino, double cantidad){
+    public boolean transferir(Long numCuentaOrigen, int cvvOrigen, Long numCuentaDestino, double cantidad) {
         return banco.transferir(numCuentaOrigen, cvvOrigen, numCuentaDestino, cantidad);
     }
